Move SQLite packages to top-level imports
Docbank needs one concise public-package convention before it adds reusable
document-understanding packages. Move the SQLite driver boundary out of pkg so
embedded applications use the same top-level convention.

This is an explicit pre-1.0 import-path break. Tagged downstream consumers can
update on their next Docbank upgrade, so no alias or forwarding package keeps
the old path alive.
