+++ internal/api/routes_backup.go
@@ -22,7 +22,7 @@ import (
 	"go.kenn.io/docbank/internal/backupapp"
 	"go.kenn.io/docbank/internal/config"
 	"go.kenn.io/docbank/internal/version"
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
+	docsqlite "go.kenn.io/docbank/sqlite"
 )
 
 type backupCreateRequest struct {
