+++ internal/backupapp/restore_placement.go
@@ -18,7 +18,7 @@ import (
 	"go.kenn.io/docbank/internal/ingest"
 	"go.kenn.io/docbank/internal/store"
 	"go.kenn.io/docbank/internal/storenamespace"
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
+	docsqlite "go.kenn.io/docbank/sqlite"
 )
 
 const (
