+++ internal/backupapp/sqlite.go
@@ -6,7 +6,7 @@ import (
 
 	"go.kenn.io/kit/backup"
 
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
+	docsqlite "go.kenn.io/docbank/sqlite"
 )
 
 type sqliteOpener struct{ driver docsqlite.Driver }
