+++ internal/blob/placement_test.go
@@ -22,7 +22,7 @@ import (
 	"go.kenn.io/docbank/internal/config"
 	"go.kenn.io/docbank/internal/jobs"
 	"go.kenn.io/docbank/internal/store"
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
+	docsqlite "go.kenn.io/docbank/sqlite"
 )
 
 func TestPlacementRunnerCopiesVerifiesAndRetiresLooseSource(t *testing.T) {
