+++ internal/store/default_driver_cgo.go
@@ -3,8 +3,8 @@
 package store
 
 import (
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
-	"go.kenn.io/docbank/pkg/sqlite/mattn"
+	docsqlite "go.kenn.io/docbank/sqlite"
+	"go.kenn.io/docbank/sqlite/mattn"
 )
 
 func defaultSQLiteDriver() docsqlite.Driver { return mattn.Driver{} }
