+++ internal/store/walk.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"sync"
 
-	docsqlite "go.kenn.io/docbank/pkg/sqlite"
+	docsqlite "go.kenn.io/docbank/sqlite"
 )
 
 const (
