+++ vault_reset_alias_unix_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.kenn.io/docbank/pkg/sqlite/modernc"
+	"go.kenn.io/docbank/sqlite/modernc"
 )
 
 func TestResetVaultRevalidatesSourceAliasAfterOwnerRelease(t *testing.T) {
